requirements.txt
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
black==23.3.0
classify-imports==4.2.0
click==8.1.3
dnspython==2.3.0
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pymongo==4.3.3
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.4
ruff==0.0.261
sniffio==1.3.0
starlette==0.26.1
tomli==2.0.1
typing-extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
